Keep the files in the checked out tree without patches.
It seems to be a good approach to have the patches applied when the package
is going to be build.

Otherwise, dpkg with the new version "3.0 (quilt)" format may become
confused (even more so if we delete the .pc directory where quilt maintains
its metadata).

We deal with an unclean tree by unpatching the sources after the build.