Use Steinar's "real" patch for IPv6. Addresses #490277, #618275, and Closes #617791.
This was taken from [Steinar H. Gunderson site][0] and had to have a little
adaptation to apply properly with libtorrent-0.12.7 (as
src/download/download_info.h was renamed to src/torrent/download_info.h).

Otherwise, it is untouched (even keeping the trailing whitespace stuff).

For the full discussion, please see:

    http://bugs.debian.org/618275

[0]: http://home.samfundet.no/~sesse/libtorrent-0.12.6-ipv6-07.patch